# be/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Money Tracker - a mobile-first PWA for family financial management.",
        "title": "Money Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@money-tracker.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/families/{id}/invites": {
            "post": {
                "security": [
                    {
                        "session": []
                    }
                ],
                "description": "Generate an invite token for family membership. Only family owner can create invites.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invite"
                ],
                "summary": "Create family invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Family ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invite created successfully",
                        "schema": {
                            "$ref": "#/definitions/model.CreateInviteResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Only family owner can create invites",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/invites/{token}/join": {
            "post": {
                "security": [
                    {
                        "session": []
                    }
                ],
                "description": "Join a family using an invite token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invite"
                ],
                "summary": "Join family with invite token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined family",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid invite token",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Token expired/used or already a member",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateInviteResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model.ErrorDetail"
                }
            }
        },
...
    }
}